requirements.txt
Flask==2.3.3
pymongo==4.6.1
python-dotenv==1.0.1
redis==6.2.0
PyJWT==2.10.1
filetype==1.2.0
pytest==8.4.1
requests==2.31.0